Refactor: InputFieldWithFixedText ausgelagert
